Docs: Add a python script to recolor UI icons to grey on transparent
There are black-on-white and black-on-transparent icon images in the
documentation. With the new dark theme for the online documentation, the
former variant looks odd, and the latter one is not visible.

The docs team chose to use switch to grey-on-transparent icons, so that
they have enough visible contrast against bright and dark backgrounds.

This change adds a python based tool that converts the existing icons to
grey-on-transparent via ImageMagick, and then optimizes the pngs via
optipng.

Task: QTCREATORBUG-26644
Change-Id: Ib14c5a68235b6b906600dcaa1c5094c9cc6ab9b7