File System view: Select root folder when clicking outside item
If the shown items are not filling the whole view, there is some empty
space at the bottom. Select the current root directory when the user
click there.

This avoids funny behavior. Without that change
- select a file or subdirectory in a directory
- collapse the parent directory
- the selected file or subdirectory stays selected, but that is not
  visible
- right-click into the empty space at the bottom
- the context menu for the not-visibly selected item is shown
- this is especially weird when e.g. choosing "New Folder"

That is not an issue if the view fills the whole space, because then a
visible item is selected when clicking.

Change-Id: I0c65e091e659c20d7b384e96ecf69c36379ed936