QtcProcess: Make API inventory
Group the common methods of ProcessInterface in the
QtcProcess API together. Order them the same.
This is a preparation step for basing
QtcProcess on ProcessInterface.

Group separately the other accessors and mutators that
base on ProcessSetupData m_setup field. This group
may became a part of ProcessInterface later.

Group the utils enhancemets (or decoration API) separately,
as that really doesn't influence the kernel of QtcProcess.

Add some TODO comments, to be considered later.

Change-Id: I0bf78284e71f4610eb0cf3508bc178f84fac679a