QtcProcess: Add a test that checks for any possible zombies
We run the tst_qtcprocess process recursively. The most
nested process blocks. After a short wait we delete
the outermost process, what initates the reaping of it.
With this test we are trying to see if reaping the middle
process reaps also its children and doesn't leave zombies.

Before we execute the test, we call Singleton::deleteAll()
in order to ensure that reaping of previously running
processes has finished. We also ensure that the number of
running tst_qtcprocess processes is zero (except of our own one).
We do this also after the test was executed.

Change-Id: I7e2bc46ad5ca22f26620da86fbaf0fa00a7db3c3