QmlDesigner: Use TypeNameString instead of Utils::SmallString
It increases the small string optimization area. Many of the type names
are longer than the SSO area of Utils::SmallString so we use a custom
String for types which has a larger SSO area. So there a much less
mallocs. This should improve performance a little but but uses a little
bit more memory.

Change-Id: I6a2225f21fb16afc8379868f3f18f38ae6ad0cd8