Utils: More forward declarations / drop unused headers [L-O]
Round 1 - focus on headers.
For classes with initial in range [L-O].
Try to keep the same separators between different kind of headers.

Drop changes to NameValueDictionary for now, as apparently
environmentfwd.h is broken currently. It looks we can't
forward declare the argument T inside QList<T> - the type
must be complete.

Change-Id: If26e88357a2ffbb91a79c4d003046443d98d5673