Squish: Fix windows toolchain for VS15
Similar to what has been done for fixing VS2017. But VS2015 also needs
to add a matching Windows Kit tools path.
This magically happens if the kit is set up manually, but for but as we
provide a stripped version of toolchains and kits we need to adjust the
kit environment manually.

Change-Id: I9a88e7424acf2768774f5ef580ef6d1fa73bcb5a