QtcProcess: Add stop() method
This method works like close() with the difference that
after calling stop() we are still going to receive
signals from QtcProcess, especially done() signal.

This method isn't blocking, so calling it is fast.
Just after calling this method the QtcProcess is still
in the same state and done() signal should be delivered
asynchronously. It's safe to call waitForFinished() after
calling stop() with the desired timeout.

Calling stop() followed by calling waitForFinished()
is a good substitute for blocking stopProcess().

Change-Id: Ie4c8fde97b2eebf9ad88750378adbd6287098a6a