QmlDesigner: Add option for smoorh rendering in form editor
Smooth rendering turns on MSAA and doubles the resolution for
rendered items. With this option enabled everything stays smooth when
zooming in. Around factor 8-10 pixels become clearly visible again, but
it still looks relativly smooth.
I added both MSAA and increasing the resolution to one option, for simplicty.
The smooth mode takes 4 times the shared memory, which should not be an issue
in most cases.
For now the option is not the default.

Task-number: QDS-7129
Task-number: QDS-7128
Change-Id: I8a778650bb40f8ba796960db9bc966e8a1efff4e