Sdktool: Remove usages of "optional"
std::optional is only supported on macOS 10.14 or later, but we build
sdktool for the installers still with Qt 5 supporting macOS 10.13.

To be able to remove our Utils::optional, which is our workaround for
macOS earlier than 10.14, we need to make sure optional is not included
by the files used by sdktool (which doesn't use any features that use
optional in any case).

Change-Id: I70341ff8c6c517087ade68eafb2d1c9c80402ec4