FilePath: Add ensureReachable()
ensureReachable(const FilePath &other) returns true if the device can access the specified path.
In case of docker devices, it will mount the specified path if possible.

The dockerdevice.cpp was split into multiple files to ease undestandability
of the parts involved.

DockerDevicePrivate was moved into its own thread and hardened for
multi thread access as many different parts of the app access it randomly.
Especially the new on-demand mounting could lead to deadlocks / race-conditions.

The DeviceShell was also hardened to better support destruction while commands
are running.

The cmake build steps are the first to implement ensureReachable calls, but more
has to be done so that other steps can also ensure reachability.

A continuing problem is posed by the way devices are cloned for editing.
This is making it impossible for configuration widgets to access internal
state of devices.
This will have to be solved before we can show users what directories are
actually mounted at any time, as we do not want to store the temporary mounts
in the devices permanently.

Change-Id: I77b64451d9b72cf63092269c75fb8ab2c7ade0ef