ProjectExplorer: Make Device lists thread-safe
Previously DeviceManager was free to access the device lists without
locking the Mutex. It was also possible that DeviceManager:: methods
would be called while the Mutex was already locked.

Moving all access to the lists into methods of DeviceManagerPrivate and
making sure that those correctly lock the Mutex fixes both issues.

Change-Id: Ia21423c37cd92ffc1318b525146d0101a66f35f1