QmlDesigner: Fix template generation for exposed properties
Since recent changes in Qt the typenames are prefixed with "<cpp>."
and "QtQuick.".
Using simplifiedTypeName() fixes this, since it removed any prefix.

Due to refactoring the check for existence of the parent property
in case of a "sub property" like was removed.

"color.red" should only be added if nothing was already added for "color".

Task-number: QDS-8856
Change-Id: I6ef6bebf0674f8499c29e20f0313b78beb2b3282