remotelinux: Don't swallow initial output
If the start of the remote process fails, e.g. because the working
directory does not exist, any error output had not been forwarded
to the QtcProcess because the pid marker had not been seen yet.

This patch changes it so that the first thing that is written is
the pid marker, so any subsequent output will arrive after it
is parsed.

In case the ssh call itself fails, the output is still forwarded
to the QtcProcess even though the marker was not parsed yet.

Change-Id: I7ee8da94cbd5113cc6f8a3a9b51fb6fe22a03a0d