Build: Simplify ts_all* targets and add "ts_all_cleaned"
Instead of trying to update all files in a single lupdate call,
make the ts_all targets just depend on the individual language targets.
It is less efficient, but not used often, simpler to setup and since
lconvert cannot work on mutliple .ts files simultanously easier to
setup a ts_all_cleaned target.

Change-Id: I60718408a661f011d84f3f74645e2f8fd49a95d1