qml2puppet: Don't always recompile qml2puppet
Whenever cmake runs, the app_version.h of qml2puppet is
recreated. This was included almost everywhere so a complete
rebuild of qml2puppet would be triggered. This patch
moves the app_version.h include into a single .cpp to limit how
much of qml2puppet needs to be recompiled.

Change-Id: I5064d450ed92ec003f164bcd42be694612011196