LocatorWidget: Reuse LocatorMatcher in LocatorWidget
Introduce QTC_USE_MATCHER env var temporarily. When
this var is set, the locator will use LocatorMatcher
internally.

It's very hard to migrate all the matchesFor() implementations
into the LocatorMatcherTask in one patch, so this var
is going to be removed when all filters are migrated.

Whenever a new patch providing the filter's migration
is added, it may be tested by setting this env var.

Change-Id: If2d6006a89f3669ea4f6f8f171e59e00917419c4