LocatorMatcherTask: Refactor the storage class
Considering the case that a locator task is running
its task in a separate thread, at least 3 threads are
involved:
1. Main thread
2. Task's thread
3. OutputFilter's thread (the thread that removes
   duplicates from tasks' results)

State before this patch:

The task (2) was always delivering the results into
the main thread (1), and when main thread received
the result it was passed to the OutputFilter's thread (3).
The drawback is that we needlessly block the main thread
for intermediate handling.

State after this patch:

The task (2) is delivering the results directly to
the OutputFilter's thread (3), without blocking
the main thread (1).
Currently the common object shared between all 3
threads is OutputDataProvider.

Rename addOutput() into reportOutput().
Make ILocatorFilter::matchers() private.

Change-Id: Ie8ad7d82ae02825e232f65b82fe329e780f58626