JavaScriptFilter: Reimplement matchers()
Add JSEngineTask class that runs JSEngine evaluator
in a separate thread. Features:
- It creates, destroys and runs evaluation of QJSEngine
  always in one, separate thread.
- In case of busy run of evaluator (e.g. because
  an input data contains a script with an endless loop)
  it's possible now to cancel the task before timeout came.
  Useful when typing fast inside locator widget.

Add also the JSEngineTaskAdapter for task tree.
The JSEngineTask is registered with JSEngine name
inside Tasking namespace.

Since QJSEngine is recreated on each search now,
the init script is always executed prior to user's input
evaluation. It lasts for about 17 ms, so that's not a
considerable overload. The benefit is that there is
no need for "Reset Engine" action now.

Change-Id: I866069c72bad17ab899f4aca9a3f6d928738f3a3