ProjectExplorer: Fix 10 seconds hang on shutdown
Delete the RunControl directly when it's not running.
Do it in two cases: when tab is closed and when
tab is reused with a new RunControl instance.

In other case, queue the call to initiateStop,
so that any possible deleteLater() is called when
the nested event loop of the shutdown is already spinning.

It looks like the amended changes introduced the hang
when compiled against Qt 6.5. When compiled against
Qt 6.4 the regression doesn't occur.
The regression is caused by
449b45ff34ce4a45e5d6a24967566f69f43ebbb6.

Amends c1f6d7044589f7842d19fc7b44574152ce72a20f
Amends 6e16512eba2c39baa714fa253843b2171ba24f47

Change-Id: Icb3ba757f32bccb3daae675b2eb516f75b84530a