QmlDesigner: Improve Model
Make node collection a little bit more explicit. Before we created,
copied and destroyed node lists. Now we send a list
around where we add nodes.

Use raw pointer for properties. Properties are used as internal pointer
plus the property name. The repesentation as an shared pointer is a
implementation detail and could be removed in the future.

Use raw pointer for internal nodes if we don't handle ownership but only
access member.

Remove double checks for properties. One lookup is better than two or
three.

Change-Id: If162f365db3e8869cbc19913abfc44d57c2d8324