ProjectExplorer: Avoid early creation of KitManager
This was only meant as part of a safety net, but in the case that
the situation this was guarding against actually happens we'd also
notice by simply not seeing the effect. This is not different to
any other place in the code where we e.g. create objects/widgets/...
and don't put them in the right container.

Change-Id: I0700fd3ca235662b718cf01229211a4eebec8a9a