PluginManager: Do not allow user interaction during delayedInit
The purpose of delayedInitialize is to have some UI that can show the
main window and potentially progress information, but user interaction
is not really supported during that phase, so prevent it.

Instead of using a timer between different plugins during
delayedInitialize, just give the event loop a chance to run.

Change-Id: I1d67da228d6358103cf8073deed5275a7ace9f3a