Utils: Restore original default DPI rounding policy
Qt Creator 11.0.2 changed the default DPI rounding policy to
"PassThrough" for Windows. In QtC 12, there is now a setting that allows
the user to chose the policy. The default was changed to "PassThrough"
also for Linux.

This change sets the default back to the original policy "Round" which
Qt Creator used from the beginnings of HighDPI support.

Amends: 3726f0d6c1d05452f070b911683c51fe95df65c4

Change-Id: I74f2310f1b5379c8dea0bf86aee965374ca2732e