QmlDesigner: Fix effect rendering
When rendering the parent of an item with layer.enabled true
and an effect we have to temporarily disable layer.effect to hide the item.

This had the side effect that the texture for the layer was deleted
and setting enabled to true was not sufficient. Resizing the item did
fix it, but using layer.live is less intrusive.
We set live to false and then back to true to ensure the texture
is rendered.

Change-Id: I1095947d768f4a18a0f8037564b1bc1289dd0813