Core: Change filterNewContent() to use a predicate
At this point this might appear slightly over-designed, but when adding
context matching and highlighting of matches the body of the lastBlock
loop will become much more complex. Something you surly don't want to
copy and paste. On the other hand the following changes are way too
complex to mix them with this refactoring. That's what this change
prepares for, and that's why it is separate.

Task-number: QTCREATORBUG-30167
Change-Id: Idb0ed2471006ebaa199acb164043f8fea8fc0d42