Doc: Small improvement of the locator pages
It was unclear what "file type" was meant. The available filters could
depend on a lot of other things, but I guess "current document and
project" is ok as a baseline.

Clarify what list of filters is meant, that is the list that appears when
the locator input field is empty. Mention the list that is shown when
selecting the magnifying glass. The descriptions in the preferences is
not longer than what appears in the tool tips, so remove that reference.

Mention the Magnifying glass > Configure menu item as an option for
opening the preference page.

Change-Id: I7d16d3f3eed11cce0aff9dcaf7e467fe0508366e