Android: support namespace in build.gradle
The use of "package" in AndroidManifest.xml has been deprecated
since AGP 7.4, and has been moved to build.gradle file instead,
set as "namespace". Qt Creator has been relying on the package
name being present in the manifest and projects not having that
will fail to deploy. This change adds support of that to Creator.

Task-number: QTBUG-106907
Change-Id: I04850a11c5840d5167f7a3d09ae76cd51dbe72ca