WIP: CppEditor: Add quickfix to show a preview of a layout
This is a toy to create a preview of a layout builder object,
triggered by Alt-Return on the constructor of a builder object.

It is limited to a very small subset of cases, essentially
only things that are practically completely non-dynanmic,
such as the example in tests/manual/layoutbuilder/v2/main.cpp.

Change-Id: I560cb3ed2aec72dd7b5afc5c96b1922b1e7fe9b2