QmlJS: Skip QtQuick3D for implicit imports
The way we resolve implicit imports creates random issues.

QtQuick3D.MaterialEditor implicitly imports QtQuick3D.

If QtQuick3D.MaterialEditor is scanned before QtQuick3D the types
of QtQuick3D are added to QtQuick3D.MaterialEditor and when later
QtQuick3D is scanned no type is added to QtQuick3D because of the cache
in ::createObjectsForImport.

The result is that QtQuick3D does not contain its types and prototypes
cannot be resolved. This does happen roughly 50% of the time and
is the reported issue.

In case QtQuick3D is scanned, first things work, but no "extra types" types are added to QtQuick3D.MaterialEditor.
To keep the patch minimal we simply skip this specific case.

QtQuick importing QtQml implicitly is not touched.

Task-number: QDS-11069
Change-Id: I74088b12a2e737a8a7467068c10a78c4be2a7bda