Chinese translation: lupdate for the translation file
Check out with submodules (`git submodule update --init --recursive`).

Configure Qt Creator:

    cmake "-DCMAKE_PREFIX_PATH=<qt_path>;<llvm_path>"
          -DCMAKE_BUILD_TYPE=Debug -G Ninja <src>

Generate ts file:

    cmake --build . --target ts_zh_CN_cleaned

Since we did a lot of changes to how the translations are
set up, this lupdate does a lot of reshuffling.
Separate it from actual translation work.
The good thing is that the changes we did will lead to
much smaller and easier to manage updates in the future.

Change-Id: I798ea80b580f4ca2aa87ca21e5d7aa6c50efa6e8