BinEditor: Paddle back a bit
Use real pointers instead of shared ones. Ensure an
editor and its document get registered correctly
inside the editor manager and the document manager.
This fixes the missing document inside the document
model as well as the missing editor which resulted
in not having a real editor at hand when using the
BinEditor.
This resulted in a missing entry inside the editor
toolbar as well as failing to close the editor
and a hard crash when trying to open another editor
in a follow-up debugging session.
We need to rethink the way of sharing the document.

Change-Id: Ia864abe8d06427ff47b96dfbffa6aa8b3c9f6ab4