QmlDesigner: Fix false dependencies
We want to get the rewriter under tests. And we want to use the designer
core in
other projects. There we cannot use plugins. So there should be no
dependency on
any plugins. There is a follow up patch which ensures, that there cannot
any
dependency on a plugin in a library.

Task-number: QDS-13406
Change-Id: I9a944fb12f7d206252946e6fe9c721c56f8536cf