SquishTests: Stabilize tst_APTW03
Enabling and disabling QML debugging in a qmake-based project causes
parsing of the project. Should issues be found during parsing, those
will overwrite previously found issues.

Wait for this parsing to finish so the expected issues found during
the following build will not be overwritten by a still ongoing parse.

Change-Id: I07665df9296e8423e281e9e1d61e5ad09aa17779