WIP: Extend Coco plugin for QMake/CMake project instrumentation
----- THIS IS A RE-BASED VERSION OF
https://codereview.qt-project.org/c/qt-creator/qt-creator/+/594659
-----

The Coco plugin now can also configure QMake and CMake projects for
the use with Coco. (COCO-1782)

* There is a new global preferences page to set the Coco
directory. (But the plugin also searches automatically for the active
Coco installation, so using the page should rarely be necessary.)

* There is a project settings page where the code coverage can be
configured, especially CoverageScanner options set.

* Code coverage is enabled by changing the build settings of a project
so that the build tool reads a special file at the
beginning (cocoplugin.prf or cocoplugin.cmake), which replaces the
normal compiler calls with calls of the Coco compiler wrappers. The
CoverageScanner options are part of this file.

* An additional, ficticious build step appears in the build menu of
QMake and CMake projects, with a button to switch coverage on or off.

* Added documentation

* The class CMakeBuildSystem has been made publicly accessible and
some member functions added so that the build settings chanbe chenged.

[ChangeLog][coco]

Added to the existing Coco plugin the ability to configure CMake and
QMake projects for code coverage.

Change-Id: I00a351b79b89bef3c25f81bb4b62ac59dc787645