Debugger: tl::expected and Utils::Result dumpers
Added dumpers for `tl::expected` and `Utils::Result` utility classes.
Added a basic test.
Useless `enum` prefix from CDB enum types (e.g. `enum MyEnum`)
will be discarded in tests (credits to @hjk).

Note that the include path for the newly added test executable
is based on `__FILE__` macro value so it might easily break
if the sources are reorganized.
Also creating a test for `Utils::Result` is practically
impossible within current setup as it is not a header only class,
so the test binary must actually link against the correct version
of Utils library target.

Fixes: QTCREATORBUG-31795
Change-Id: I7f9ccb92d0c59333a2dca4ba928ac991f1e5238b