Android: optionally add a default system image for an AVD
Find out the highest API level that Qt versions are using,
and add a Google Play Store Emulator image to the list of
packages to install only if there are missing essential
packages. This can practically work as an optional package
so that it be most suitable for users doing first time setup
but if the essential packages are all installed, then don't
invalidate the settings.

Fixes: QTCREATORBUG-24754
Task-number: QTCREATORBUG-31837
Change-Id: Ia1911637a846eb2963d69db6d33346dc04dc16b6