WIP: Add checks without linking
AstCheck is a slimmed down version doing checks direcrtly on the AST.
We need this because certain things will not be reprensented in the model or
are already invalid in the model (e.g. invalid ids)

ModelCheck does use the model. There is still an issue with retrieving the line number.
Many checks are simpler on the model and should be ported.

Change-Id: I753cbc259b21129863db54926f77a07391224b2d