ProjectExplorer: Introduce builddevice and rundevice FilePath schemes
This is meant to be used to conveniently refer to device path
before the concrete instance is known and without relying
on implicit assumption on what an unqualified name could mean.

Especially in remote build setups, all host, build and run device
are legitimate locations for debugger binaries.

These paths can be resolved as soon as a kit is known.

Change-Id: Ia8b8a90ad6298689d5715525d918cd25894172a8