CMakeParser: Get rid of the std::optional holding a container
There is no need to enclose a container inside the std::optional,
as if the container is empty, it indicates that no values are in.
Holding optional for a container class may lead to the confusing
behavior, that the optional itself isn't empty, but holding an
empty container.

Instead, introduce a m_callStackDetected flag.

Task-number: QTCREATORBUG-29965
Change-Id: If8d43e0633286c076e3da436eb61114aa68f4c35