Change the section header for cherry-pick bot to work around Fixes bot's behavior.
Qt-Creator is not truly in cherry-pick mode, as forward-merge remains
the primary workflow, but a request in May of 2024 requested cherry-pick automation to be enabled in the repo regardless. This
had the side effect of causing the Fixes bot to fail in finding the next
version due to internal flags for if a repo is in cherry-pick or
forward-merge mode.

See commit: ba65e52cefa91f71b2fd51c395db609049032d53

This change works around this behavior issue by altering the name of the section header that the Fixes bot is looking for in the webhooks.config file, effectively allowing the Fixes bot to correctly determine that the repo remains in forward merge mode.

Change-Id: I57335576b5f4943fbd9a7420376f168f96512f74